requirements.txt
numpy
scipy
matplotlib 
pyserial
scapy
paho-mqtt
pandas